fix(gui-app): satisfy lint for the shared image action bar
Split the pure helpers out of the image-actions component module so
react-refresh sees a components-only file, and stop typing the test's
render helper as ReactNode - React 19 folds a promise into that union,
which reads to no-floating-promises as an unawaited call.
